+++ client/client_apis.go
@@ -1,3 +1,4 @@
+// Package client provides a Go client for interacting with the Accelerated Cloud Storage service.
 package client
 
 import (
@@ -17,6 +18,8 @@ import (
 )
 
 // CreateBucket sends a request to create a new bucket.
+// It requires a bucket name and region specification.
+// Returns an error if the bucket creation fails.
 func (client *ACSClient) CreateBucket(ctx context.Context, bucket, region string) error {
 	req := &pb.CreateBucketRequest{
 		Bucket: bucket,
@@ -32,6 +35,7 @@ func (client *ACSClient) CreateBucket(ctx context.Context, bucket, region string
 }
 
 // DeleteBucket requests deletion of the specified bucket.
+// Returns an error if the bucket deletion fails or the bucket doesn't exist.
 func (client *ACSClient) DeleteBucket(ctx context.Context, bucket string) error {
 	req := &pb.DeleteBucketRequest{
 		Bucket: bucket,
@@ -46,6 +50,7 @@ func (client *ACSClient) DeleteBucket(ctx context.Context, bucket string) error 
 }
 
 // ListBuckets retrieves all buckets from the server.
+// Returns a list of bucket objects and an error if the operation fails.
 func (client *ACSClient) ListBuckets(ctx context.Context) ([]*pb.Bucket, error) {
 	req := &pb.ListBucketsRequest{}
 
@@ -58,9 +63,11 @@ func (client *ACSClient) ListBuckets(ctx context.Context) ([]*pb.Bucket, error) 
 }
 
 // PutObject uploads data to the specified bucket and key.
+// It automatically handles compression for large objects when beneficial.
+// Returns an error if the upload fails.
 func (client *ACSClient) PutObject(ctx context.Context, bucket, key string, data []byte) error {
 	// Only compress if data is larger than threshold and compression would be beneficial
-	const compressionThreshold = 100 * 1024 * 1024 // 1MB threshold 
+	const compressionThreshold = 100 * 1024 * 1024 // 1MB threshold
 	isCompressed := false
 	if len(data) >= compressionThreshold {
 		var buf bytes.Buffer
@@ -73,8 +80,8 @@ func (client *ACSClient) PutObject(ctx context.Context, bucket, key string, data
 		}
 		if err := gw.Close(); err != nil {
 			return fmt.Errorf("failed to close gzip writer: %v", err)
-			}
-		
+		}
+
 		// Only use compression if it actually reduces size
 		if buf.Len() < len(data) {
 			isCompressed = true
@@ -128,6 +135,7 @@ func (client *ACSClient) PutObject(ctx context.Context, bucket, key string, data
 }
 
 // GetObject downloads the specified object from the server.
+// Returns the object's data and an error if the download fails.
 func (client *ACSClient) GetObject(ctx context.Context, bucket, key string) ([]byte, error) {
 	req := &pb.GetObjectRequest{
 		Bucket: bucket,
@@ -165,6 +173,7 @@ func (client *ACSClient) GetObject(ctx context.Context, bucket, key string) ([]b
 }
 
 // DeleteObject removes a single object from a bucket.
+// Returns an error if the deletion fails.
 func (client *ACSClient) DeleteObject(ctx context.Context, bucket, key string) error {
 	req := &pb.DeleteObjectRequest{
 		Bucket: bucket,
@@ -180,6 +189,7 @@ func (client *ACSClient) DeleteObject(ctx context.Context, bucket, key string) e
 }
 
 // HeadObject retrieves metadata for a specific object.
+// Returns the object's metadata and an error if the operation fails.
 func (client *ACSClient) HeadObject(ctx context.Context, bucket, key string) (*HeadObjectOutput, error) {
 	req := &pb.HeadObjectRequest{
 		Bucket: bucket,
@@ -192,19 +202,20 @@ func (client *ACSClient) HeadObject(ctx context.Context, bucket, key string) (*H
 	}
 
 	return &HeadObjectOutput{
-		ContentType:     resp.Metadata.ContentType,
-		ContentLength:   resp.Metadata.Size,
-		LastModified:    resp.Metadata.LastModified.AsTime(),
-		ETag:            resp.Metadata.Etag,
-		ContentEncoding: resp.Metadata.ContentEncoding,
-		ContentLanguage: resp.Metadata.ContentLanguage,
-		VersionId:       resp.Metadata.VersionId,
+		ContentType:          resp.Metadata.ContentType,
+		ContentLength:        resp.Metadata.Size,
+		LastModified:         resp.Metadata.LastModified.AsTime(),
+		ETag:                 resp.Metadata.Etag,
+		ContentEncoding:      resp.Metadata.ContentEncoding,
+		ContentLanguage:      resp.Metadata.ContentLanguage,
+		VersionId:            resp.Metadata.VersionId,
 		ServerSideEncryption: resp.Metadata.ServerSideEncryption,
-		UserMetadata:    resp.Metadata.UserMetadata,
+		UserMetadata:         resp.Metadata.UserMetadata,
 	}, nil
 }
 
 // DeleteObjects requests bulk deletion of objects in a bucket.
+// Returns an error if any object deletion fails.
 func (client *ACSClient) DeleteObjects(bucket string, keys []string) error {
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
@@ -234,6 +245,7 @@ func (client *ACSClient) DeleteObjects(bucket string, keys []string) error {
 }
 
 // ListObjects retrieves object keys from the server based on given options.
+// Returns a list of object keys and an error if the operation fails.
 func (client *ACSClient) ListObjects(ctx context.Context, bucket string, opts *ListObjectsOptions) ([]string, error) {
 	req := &pb.ListObjectsRequest{
 		Bucket: bucket,
@@ -274,64 +286,66 @@ func (client *ACSClient) ListObjects(ctx context.Context, bucket string, opts *L
 }
 
 // HeadBucket retrieves metadata for a specific bucket.
+// Returns the bucket's metadata and an error if the operation fails.
 func (client *ACSClient) HeadBucket(ctx context.Context, bucket string) (*HeadBucketOutput, error) {
-    req := &pb.HeadBucketRequest{
-        Bucket: bucket,
-    }
+	req := &pb.HeadBucketRequest{
+		Bucket: bucket,
+	}
 
-    resp, err := client.client.HeadBucket(ctx, req)
-    if err != nil {
-        return nil, fmt.Errorf("failed to head bucket: %v", err)
-    }
+	resp, err := client.client.HeadBucket(ctx, req)
+	if err != nil {
+		return nil, fmt.Errorf("failed to head bucket: %v", err)
+	}
 
-    return &HeadBucketOutput{
-        Region: resp.BucketRegion,
-
-    }, nil
+	return &HeadBucketOutput{
+		Region: resp.BucketRegion,
+	}, nil
 }
 
-// RotateKey checks if key rotation is needed and performs it if necessary
+// RotateKey checks if key rotation is needed and performs it if necessary.
+// The force parameter can be used to require rotation regardless of timing.
+// Returns an error if the rotation fails.
 func (client *ACSClient) RotateKey(ctx context.Context, force bool) error {
-    // Get the home directory
-    homeDir, err := os.UserHomeDir()
-    if err != nil {
-        return fmt.Errorf("failed to get home directory: %v", err)
-    }
+	// Get the home directory
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		return fmt.Errorf("failed to get home directory: %v", err)
+	}
 
-    credsFile := filepath.Join(homeDir, ".acs", "credentials.yaml")
-    
-    // Read the entire credentials file
-    data, err := os.ReadFile(credsFile)
-    if err != nil {
-        return fmt.Errorf("failed to read credentials file: %v", err)
-    }
+	credsFile := filepath.Join(homeDir, ".acs", "credentials.yaml")
 
-    var profiles profileCredentials
-    if err := yaml.Unmarshal(data, &profiles); err != nil {
-        return fmt.Errorf("failed to unmarshal credentials: %v", err)
-    }
+	// Read the entire credentials file
+	data, err := os.ReadFile(credsFile)
+	if err != nil {
+		return fmt.Errorf("failed to read credentials file: %v", err)
+	}
 
-    // Get current profile name from environment variable
-    profile := os.Getenv("ACS_PROFILE")
-    if profile == "" {
-        profile = "default"
-    }
+	var profiles profileCredentials
+	if err := yaml.Unmarshal(data, &profiles); err != nil {
+		return fmt.Errorf("failed to unmarshal credentials: %v", err)
+	}
 
-    creds, ok := profiles[profile]
-    if (!ok) {
-        return fmt.Errorf("profile '%s' not found in credentials file", profile)
-    }
+	// Get current profile name from environment variable
+	profile := os.Getenv("ACS_PROFILE")
+	if profile == "" {
+		profile = "default"
+	}
 
-    resp, err := client.client.RotateKey(ctx, &pb.RotateKeyRequest{
-        AccessKeyId: creds.AccessKeyID,
-        Force:      &force,
-    })
-    if err != nil {
-        return fmt.Errorf("key rotation failed: %v", err)
-    }
-    if !resp.Rotated {
-        return nil
-    }
+	creds, ok := profiles[profile]
+	if !ok {
+		return fmt.Errorf("profile '%s' not found in credentials file", profile)
+	}
+
+	resp, err := client.client.RotateKey(ctx, &pb.RotateKeyRequest{
+		AccessKeyId: creds.AccessKeyID,
+		Force:       &force,
+	})
+	if err != nil {
+		return fmt.Errorf("key rotation failed: %v", err)
+	}
+	if !resp.Rotated {
+		return nil
+	}
 
 	// Update only the current profile's credentials
 	updatedCreds := creds
@@ -339,58 +353,59 @@ func (client *ACSClient) RotateKey(ctx context.Context, force bool) error {
 	profiles[profile] = updatedCreds
 
 	// Marshal updated profiles
-    data, err = yaml.Marshal(profiles)
-    if err != nil {
-        return fmt.Errorf("failed to marshal credentials: %v", err)
-    }
+	data, err = yaml.Marshal(profiles)
+	if err != nil {
+		return fmt.Errorf("failed to marshal credentials: %v", err)
+	}
 
-    // Write back to file
-    if err := os.WriteFile(credsFile, data, 0600); err != nil {
-        return fmt.Errorf("failed to update credentials file: %v", err)
-    }
+	// Write back to file
+	if err := os.WriteFile(credsFile, data, 0600); err != nil {
+		return fmt.Errorf("failed to update credentials file: %v", err)
+	}
 
-    return nil
+	return nil
 }
 
-// ShareBucket informs the service about a bucket that has been shared with it
+// ShareBucket informs the service about a bucket that has been shared with it.
+// Returns an error if the sharing operation fails or permissions are insufficient.
 func (client *ACSClient) ShareBucket(ctx context.Context, bucket string) error {
-    req := &pb.ShareBucketRequest{
-        BucketName: bucket,
-    }
+	req := &pb.ShareBucketRequest{
+		BucketName: bucket,
+	}
 
-    _, err := client.client.ShareBucket(ctx, req)
-    if err != nil {
-        // Check for specific error types and provide clear messages
-        st, ok := status.FromError(err)
-        if ok {
-            switch st.Code() {
-            case codes.NotFound:
-                return fmt.Errorf("bucket does not exist: %v", err)
-            case codes.PermissionDenied:
-                return fmt.Errorf("service lacks permission to access bucket: %v", err)
-            case codes.InvalidArgument:
-                return fmt.Errorf("invalid bucket name: %v", err)
-            }
-        }
-        return fmt.Errorf("failed to share bucket: %v", err)
-    }
+	_, err := client.client.ShareBucket(ctx, req)
+	if err != nil {
+		// Check for specific error types and provide clear messages
+		st, ok := status.FromError(err)
+		if ok {
+			switch st.Code() {
+			case codes.NotFound:
+				return fmt.Errorf("bucket does not exist: %v", err)
+			case codes.PermissionDenied:
+				return fmt.Errorf("service lacks permission to access bucket: %v", err)
+			case codes.InvalidArgument:
+				return fmt.Errorf("invalid bucket name: %v", err)
+			}
+		}
+		return fmt.Errorf("failed to share bucket: %v", err)
+	}
 
-    return nil
+	return nil
 }
 
 // CopyObject copies an object from a source bucket/key to a destination bucket/key.
+// Returns an error if the copy operation fails.
 func (client *ACSClient) CopyObject(ctx context.Context, bucket, copySource, key string) error {
-    req := &pb.CopyObjectRequest{
+	req := &pb.CopyObjectRequest{
 		Bucket:     bucket,
 		CopySource: copySource,
 		Key:        key,
 	}
 
-    _, err := client.client.CopyObject(ctx, req)
-    if err != nil {
-        return fmt.Errorf("failed to copy object: %v", err)
-    }
+	_, err := client.client.CopyObject(ctx, req)
+	if err != nil {
+		return fmt.Errorf("failed to copy object: %v", err)
+	}
 
-    return nil
+	return nil
 }
-
